+++ CasualTasker/Infrastructure/Middleware/IExceptionHandlingService.cs
@@ -1,28 +1,14 @@
-﻿using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CasualTasker.Infrastructure.Middleware
+﻿namespace CasualTasker.Infrastructure.Middleware
 {
+    /// <summary>
+    /// Interface for handling exceptions.
+    /// </summary>
     public interface IExceptionHandlingService
     {
+        /// <summary>
+        /// Handles the specified exception.
+        /// </summary>
+        /// <param name="exception">The exception to handle.</param>
         void HandleException(Exception exception);
-    }
-
-    public class ExceptionHandlingService : IExceptionHandlingService
-    {
-        private readonly ILogger<ExceptionHandlingService> _logger;
-        public ExceptionHandlingService(ILogger<ExceptionHandlingService> logger)
-        {
-            _logger = logger;
-        }
-
-        public void HandleException(Exception exception)
-        {
-            _logger.LogError(exception, exception.Message);
-        }
     }
 }
